remove hadron-react-utils package
