+++ packages/hadron-react-bson/package.json
@@ -37,7 +37,6 @@
     "react-dom": "^16.8.0"
   },
   "dependencies": {
-    "hadron-react-utils": "^5.3.0",
     "hex-to-uuid": "^1.0.1",
     "moment-timezone": "^0.5.14"
   },
