+++ packages/hadron-react-bson/test/truncate.test.js
@@ -1,0 +1,20 @@
+const { expect } = require('chai');
+const truncate = require('../src/truncate');
+
+describe('#truncate', () => {
+  context('when a limit is provided', () => {
+    const value = truncate('testing', 3);
+
+    it('truncates the string at the limit and adds elipsis', () => {
+      expect(value).to.equal('tes...');
+    });
+  });
+
+  context('when a limit is not provided', () => {
+    const value = truncate('testing');
+
+    it('does not truncate the string', () => {
+      expect(value).to.equal('testing');
+    });
+  });
+});
