--- packages/hadron-react-utils/.eslintignore
@@ -1,1 +1,0 @@
-lib
