--- packages/hadron-react-utils/.gitignore
@@ -1,1 +1,0 @@
-lib
