--- packages/hadron-react-utils/.npmignore
@@ -1,11 +1,0 @@
-.babelrc
-.eslintignore
-.eslintrc
-config/
-mocha-webpack.opts
-.jshintrc
-.jsfmtrc
-.travis.yml
-.zuul.yml
-src/
-test/
