--- packages/hadron-react-utils/README.md
@@ -1,16 +1,0 @@
-# hadron-react-utils [![][npm_img]][npm_url]
-
-> Hadron React Utilities
-
-### Usage
-
-Utility function for truncating strings with ellipsis.
-
-```javascript
-const { truncate } = require('hadron-react-utils');
-
-tuncate('testing', 3);
-```
-
-[npm_img]: https://img.shields.io/npm/v/hadron-react-utils.svg?style=flat-square
-[npm_url]: https://www.npmjs.org/package/hadron-react-utils
