--- packages/hadron-react-utils/config/webpack.base.config.js
@@ -1,26 +1,0 @@
-const path = require('path');
-
-const config = {
-  mode: process.env.NODE_ENV !== 'production' ? 'development' : 'production',
-  resolve: {
-    modules: ['node_modules'],
-    extensions: ['.js', '.jsx']
-  },
-  module: {
-    rules: [
-      {
-        test: /\.(js|jsx)$/,
-        use: [{
-          loader: 'babel-loader',
-          options: {
-            root: path.resolve(__dirname, '..'),
-            cacheDirectory: !process.env.CI
-          }
-        }],
-        exclude: /(node_modules)/
-      }
-    ]
-  }
-};
-
-module.exports = config;
