--- packages/hadron-react-utils/mocha-webpack.opts
@@ -1,5 +1,0 @@
---webpack-config config/webpack.test.config.js
---require jsdom-global/register
---recursive
---ui bdd
---timeout 15000
