--- packages/hadron-react-utils/package.json
@@ -1,68 +1,0 @@
-{
-  "name": "hadron-react-utils",
-  "description": "Hadron React Utils",
-  "author": "Durran Jordan <durran@gmail.com>",
-  "bugs": {
-    "url": "https://jira.mongodb.org/projects/COMPASS/issues",
-    "email": "compass@mongodb.com"
-  },
-  "homepage": "https://github.com/mongodb-js/compass",
-  "version": "5.3.0",
-  "main": "lib/index.js",
-  "module": "lib/index.js",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/mongodb-js/compass.git"
-  },
-  "license": "SSPL",
-  "keywords": [
-    "mongodb-js"
-  ],
-  "scripts": {
-    "clean": "rimraf lib",
-    "precompile": "npm run clean",
-    "compile": "cross-env NODE_ENV=production webpack --config ./config/webpack.prod.config.js",
-    "pretest": "npm run compile",
-    "test": "cross-env NODE_ENV=test mocha-webpack \"./test/**/*.test.js\"",
-    "lint": "eslint \"./{src,lib,test,bin}/**/*.{js,jsx,ts,tsx}\" \"./*.js\" --no-error-on-unmatched-pattern",
-    "depcheck": "depcheck",
-    "check": "npm run lint && npm run depcheck",
-    "test-ci": "npm run test",
-    "prepublishOnly": "npm run compile",
-    "bootstrap": "npm run compile"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.14.3",
-    "@babel/plugin-proposal-decorators": "^7.14.2",
-    "@babel/plugin-syntax-dynamic-import": "^7.8.3",
-    "@babel/preset-env": "^7.14.2",
-    "@babel/preset-react": "^7.13.13",
-    "@babel/register": "^7.13.16",
-    "babel-loader": "^8.2.2",
-    "chai": "^4.1.2",
-    "chai-enzyme": "1.0.0-beta.0",
-    "core-js": "^3.12.1",
-    "cross-env": "^7.0.0",
-    "depcheck": "^1.4.0",
-    "enzyme": "^3.7.0",
-    "enzyme-adapter-react-16": "^1.6.0",
-    "eslint": "^7.25.0",
-    "eslint-config-mongodb-js": "^5.0.3",
-    "ignore-loader": "^0.1.2",
-    "jsdom": "^16.6.0",
-    "jsdom-global": "^3.0.2",
-    "mocha": "^7.1.0",
-    "mocha-webpack": "^2.0.0-beta.0",
-    "peer-deps-externals-webpack-plugin": "^1.0.4",
-    "rimraf": "^3.0.2",
-    "sinon": "^2.1.0",
-    "webpack": "^4.46.0",
-    "webpack-cli": "^3.3.12",
-    "webpack-merge": "^4.2.2",
-    "webpack-node-externals": "^3.0.0"
-  },
-  "dependencies": {
-    "react": "^16.14.0",
-    "react-dom": "^16.14.0"
-  }
-}
