--- packages/hadron-react-utils/src/index.js
@@ -1,1 +1,0 @@
-module.exports.truncate = require('./truncate');
